requirements.txt
# Backend
fastapi
uvicorn
requests
python-dotenv
pydantic

# Frontend
streamlit

# Voice Input (Google ASR)
SpeechRecognition
pyaudio

# Voice Output (TTS)
pyttsx3
pypiwin32
pywin32

# Install
# pip install -r requirements.txt

# Run Backend
# uvicorn backend.main:app --reload

# Run Frontend
# streamlit run frontend/app.py
